Fix compilation errors and complete test suite implementation
- Update solution file to include new Application and API projects
- Fix domain entity constructors and method signatures for API compatibility
- Resolve namespace inconsistencies across Infrastructure layer
- Update repository interfaces and implementations for proper nullable types
- Fix BPM type conversions throughout test files (int to ushort casting)
- Simplify PlayerScores API controller to reduce complexity
- Add comprehensive test coverage for PlayerScoreService and PlayerScoresController
- Ensure all projects build successfully with complete test suite

+++ src/API/Controllers/MusicController.cs
@@ -54,7 +54,7 @@ public class MusicController : ControllerBase
         var title = new MusicTitle(request.Title);
         var artist = new Artist(request.Artist);
         var genre = new Genre(request.Genre);
-        var bpm = new Bpm(request.Bpm);
+        var bpm = new Bpm((ushort)request.Bpm);
         
         var music = await _musicService.CreateMusicAsync(title, artist, genre, bpm);
         
@@ -87,7 +87,7 @@ public class MusicController : ControllerBase
             music.ChangeGenre(new Genre(request.Genre));
         
         if (request.Bpm.HasValue)
-            music.ChangeBpm(new Bpm(request.Bpm.Value));
+            music.ChangeBpm(new Bpm((ushort)request.Bpm.Value));
 
         await _musicService.UpdateMusicAsync(music);
