+++ tests/DomainTest/ValueObjectTests/BpmTests.cs
@@ -8,14 +8,14 @@ namespace DomainTest.ValueObjectTests
         [Fact]
         public void Bpm_MustBeGreaterThanZero()
         {
-            var exception = Assert.Throws<ArgumentException>(() => new Bpm(0));
+            var exception = Assert.Throws<ArgumentException>(() => new Bpm((ushort)0));
             Assert.Equal("Bpm must be greater than 0.", exception.Message);
         }
 
         [Fact]
         public void Bpm_CanBeCreated_WithValidValue()
         {
-            var bpm = new Bpm(120);
+            var bpm = new Bpm((ushort)120);
             Assert.Equal(120, bpm.Value);
         }
     }
